chore: update tutor version to 19.0.2 (#4)
The openedx-auth patch was moved to the bottom of the auth.yml file in tutor v19.0.2 in https://github.com/overhangio/tutor/pull/1191. 

This change is necessary to allow patching the database setting for setting postgresql as the main database through the openedx-auth used by openedx to connect to the database.